from and processor functions moved to build-utils + code and tests are still broken + absolute path removed from  functions
